@@ -1,5 +1,4 @@
 import React from "react"
-import { createStyles, makeStyles, Theme } from "@material-ui/core/styles"
 import {
   Table,
   TableBody,
@@ -11,41 +10,19 @@ import {
   TableSortLabel,
   Collapse,
   IconButton,
-} from "@material-ui/core"
-import KeyboardArrowDownIcon from "@material-ui/icons/KeyboardArrowDown"
-import KeyboardArrowUpIcon from "@material-ui/icons/KeyboardArrowUp"
-import { Clear } from "@material-ui/icons"
+  useTheme,
+} from "@mui/material"
+import { styled } from "@mui/system"
+import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown"
+import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp"
+import { Clear } from "@mui/icons-material"
 
 type Order = "asc" | "desc"
 
-const defaultRowsPerPageOption = [10, 50, 100, { label: "All", value: -1 }]
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+type Value = any
 
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    root: {
-      width: "100%",
-    },
-    visuallyHidden: {
-      border: 0,
-      clip: "rect(0 0 0 0)",
-      height: 1,
-      margin: -1,
-      overflow: "hidden",
-      padding: 0,
-      position: "absolute",
-      top: 20,
-      width: 1,
-    },
-    filterable: {
-      color: theme.palette.primary.main,
-      textDecoration: "underline",
-      cursor: "pointer",
-    },
-    tableHeaderCell: {
-      display: "inline-flex",
-    },
-  })
-)
+const defaultRowsPerPageOption = [10, 50, 100, { label: "All", value: -1 }]
 
 interface DataGridColumn<T> {
   field: keyof T
@@ -54,12 +31,12 @@ interface DataGridColumn<T> {
   less?: (a: T, b: T) => number
   filterable?: boolean
   toCellValue?: (rowIndex: number) => string | React.ReactNode
-  padding?: "default" | "checkbox" | "none"
+  padding?: "normal" | "checkbox" | "none"
 }
 
-interface RowFilter<T> {
-  field: keyof T
-  value: any
+interface RowFilter {
+  columnIdx: number
+  value: Value
 }
 
 function DataGrid<T>(props: {
@@ -70,14 +47,14 @@ function DataGrid<T>(props: {
   collapseBody?: (rowIndex: number) => React.ReactNode
   initialRowsPerPage?: number
   rowsPerPageOption?: Array<number | { value: number; label: string }>
-}) {
-  const classes = useStyles()
-  const { columns, rows, keyField, dense, collapseBody } = props
+  defaultFilter?: (row: T) => boolean
+}): React.ReactElement {
+  const { columns, rows, keyField, dense, collapseBody, defaultFilter } = props
   let { initialRowsPerPage, rowsPerPageOption } = props
   const [order, setOrder] = React.useState<Order>("asc")
   const [orderBy, setOrderBy] = React.useState<number>(0) // index of columns
   const [page, setPage] = React.useState(0)
-  const [filters, setFilters] = React.useState<RowFilter<T>[]>([])
+  const [filters, setFilters] = React.useState<RowFilter[]>([])
 
   const getRowIndex = (row: T): number => {
     return rows.findIndex((row2) => row[keyField] === row2[keyField])
@@ -104,33 +81,45 @@ function DataGrid<T>(props: {
   }
 
   // Filtering
-  const fieldAlreadyFiltered = (field: keyof T): boolean =>
-    filters.some((f) => f.field === field)
+  const fieldAlreadyFiltered = (columnIdx: number): boolean =>
+    filters.some((f) => f.columnIdx === columnIdx)
 
-  const handleClickFilterCell = (field: keyof T, value: any) => {
-    if (fieldAlreadyFiltered(field)) {
+  const handleClickFilterCell = (columnIdx: number, value: Value) => {
+    if (fieldAlreadyFiltered(columnIdx)) {
       return
     }
-    const newFilters = [...filters, { field: field, value: value }]
+    const newFilters = [...filters, { columnIdx: columnIdx, value: value }]
     setFilters(newFilters)
   }
 
-  const clearFilter = (field: keyof T): void => {
-    setFilters(filters.filter((f) => f.field !== field))
+  const clearFilter = (columnIdx: number): void => {
+    setFilters(filters.filter((f) => f.columnIdx !== columnIdx))
   }
 
-  const filteredRows = rows.filter((row) =>
-    filters.length === 0
+  const filteredRows = rows.filter((row, rowIdx) => {
+    if (defaultFilter !== undefined && defaultFilter(row)) {
+      return false
+    }
+    return filters.length === 0
       ? true
       : filters.some((f) => {
-          return row[f.field] === f.value
+          if (columns.length <= f.columnIdx) {
+            console.log(
+              `columnIdx=${f.columnIdx} must be smaller than columns.length=${columns.length}`
+            )
+            return true
+          }
+          const toCellValue = columns[f.columnIdx].toCellValue
+          if (toCellValue !== undefined) {
+            return toCellValue(rowIdx) === f.value
+          }
+          const field = columns[f.columnIdx].field
+          return row[field] === f.value
         })
-  )
+  })
 
   // Sorting
-  const createSortHandler = (columnId: number) => (
-    event: React.MouseEvent<unknown>
-  ) => {
+  const createSortHandler = (columnId: number) => () => {
     const isAsc = orderBy === columnId && order === "asc"
     setOrder(isAsc ? "desc" : "asc")
     setOrderBy(columnId)
@@ -143,8 +132,25 @@ function DataGrid<T>(props: {
   const emptyRows =
     rowsPerPage - Math.min(rowsPerPage, sortedRows.length - page * rowsPerPage)
 
+  const RootDiv = styled("div")({
+    width: "100%",
+  })
+  const HiddenSpan = styled("span")({
+    border: 0,
+    clip: "rect(0 0 0 0)",
+    height: 1,
+    margin: -1,
+    overflow: "hidden",
+    padding: 0,
+    position: "absolute",
+    top: 20,
+    width: 1,
+  })
+  const TableHeaderCellSpan = styled("span")({
+    display: "inline-flex",
+  })
   return (
-    <div className={classes.root}>
+    <RootDiv>
       <TableContainer>
         <Table
           aria-labelledby="tableTitle"
@@ -154,26 +160,26 @@ function DataGrid<T>(props: {
           <TableHead>
             <TableRow>
               {collapseBody ? <TableCell /> : null}
-              {columns.map((column, index) => (
+              {columns.map((column, columnIdx) => (
                 <TableCell
-                  key={index}
-                  padding={column.padding || "default"}
+                  key={columnIdx}
+                  padding={column.padding || "normal"}
                   sortDirection={orderBy === column.field ? order : false}
                 >
-                  <span className={classes.tableHeaderCell}>
+                  <TableHeaderCellSpan>
                     {column.sortable ? (
                       <TableSortLabel
-                        active={orderBy === index}
-                        direction={orderBy === index ? order : "asc"}
-                        onClick={createSortHandler(index)}
+                        active={orderBy === columnIdx}
+                        direction={orderBy === columnIdx ? order : "asc"}
+                        onClick={createSortHandler(columnIdx)}
                       >
                         {column.label}
                         {orderBy === column.field ? (
-                          <span className={classes.visuallyHidden}>
+                          <HiddenSpan>
                             {order === "desc"
                               ? "sorted descending"
                               : "sorted ascending"}
-                          </span>
+                          </HiddenSpan>
                         ) : null}
                       </TableSortLabel>
                     ) : (
@@ -183,25 +189,25 @@ function DataGrid<T>(props: {
                       <IconButton
                         size={dense ? "small" : "medium"}
                         style={
-                          fieldAlreadyFiltered(column.field)
+                          fieldAlreadyFiltered(columnIdx)
                             ? {}
                             : { visibility: "hidden" }
                         }
                         color="inherit"
-                        onClick={(e) => {
-                          clearFilter(column.field)
+                        onClick={() => {
+                          clearFilter(columnIdx)
                         }}
                       >
                         <Clear />
                       </IconButton>
                     ) : null}
-                  </span>
+                  </TableHeaderCellSpan>
                 </TableCell>
               ))}
             </TableRow>
           </TableHead>
           <TableBody>
-            {currentPageRows.map((row, index) => (
+            {currentPageRows.map((row) => (
               <DataGridRow<T>
                 columns={columns}
                 rowIndex={getRowIndex(row)}
@@ -226,10 +232,10 @@ function DataGrid<T>(props: {
         count={filteredRows.length}
         rowsPerPage={rowsPerPage}
         page={page}
-        onChangePage={handleChangePage}
-        onChangeRowsPerPage={handleChangeRowsPerPage}
+        onPageChange={handleChangePage}
+        onRowsPerPageChange={handleChangeRowsPerPage}
       />
-    </div>
+    </RootDiv>
   )
 }
 
@@ -239,9 +245,8 @@ function DataGridRow<T>(props: {
   row: T
   keyField: keyof T
   collapseBody?: (rowIndex: number) => React.ReactNode
-  handleClickFilterCell: (field: keyof T, value: any) => void
+  handleClickFilterCell: (columnIdx: number, value: Value) => void
 }) {
-  const classes = useStyles()
   const {
     columns,
     rowIndex,
@@ -251,7 +256,13 @@ function DataGridRow<T>(props: {
     handleClickFilterCell,
   } = props
   const [open, setOpen] = React.useState(false)
+  const theme = useTheme()
 
+  const FilterableDiv = styled("div")({
+    color: theme.palette.primary.main,
+    textDecoration: "underline",
+    cursor: "pointer",
+  })
   return (
     <React.Fragment>
       <TableRow hover tabIndex={-1}>
@@ -269,22 +280,27 @@ function DataGridRow<T>(props: {
         {columns.map((column, columnIndex) => {
           const cellItem = column.toCellValue
             ? column.toCellValue(rowIndex)
-            : row[column.field]
+            : // TODO(c-bata): Avoid this implicit type conversion.
+              (row[column.field] as number | string | null | undefined)
 
           return column.filterable ? (
             <TableCell
-              key={`${row[keyField]}:${column.field}:${columnIndex}`}
-              padding={column.padding || "default"}
-              onClick={(e) => {
-                handleClickFilterCell(column.field, row[column.field])
+              key={`${row[keyField]}:${column.field.toString()}:${columnIndex}`}
+              padding={column.padding || "normal"}
+              onClick={() => {
+                const value =
+                  column.toCellValue !== undefined
+                    ? column.toCellValue(rowIndex)
+                    : row[column.field]
+                handleClickFilterCell(columnIndex, value)
               }}
             >
-              <div className={classes.filterable}>{cellItem}</div>
+              <FilterableDiv>{cellItem}</FilterableDiv>
             </TableCell>
           ) : (
             <TableCell
-              key={`${row[keyField]}:${column.field}:${columnIndex}`}
-              padding={column.padding || "default"}
+              key={`${row[keyField]}:${column.field.toString()}:${columnIndex}`}
+              padding={column.padding || "normal"}
             >
               {cellItem}
             </TableCell>
