@@ -1,306 +1,240 @@
-import React, { FC, useEffect } from "react"
+import React, { FC, useEffect, useMemo, useState } from "react"
+import { useNavigate, useLocation } from "react-router-dom"
 import { useRecoilValue } from "recoil"
 import { Link } from "react-router-dom"
-import { createStyles, makeStyles, Theme } from "@material-ui/core/styles"
 import {
-  AppBar,
-  Toolbar,
   Typography,
   Container,
   Card,
-  Grid,
+  CardActionArea,
   Box,
   Button,
   IconButton,
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  DialogContentText,
+  MenuItem,
+  useTheme,
+  InputAdornment,
+  SvgIcon,
+  CardContent,
   TextField,
-  DialogActions,
-  FormControlLabel,
-  Checkbox,
-} from "@material-ui/core"
-import { AddBox, Delete, Refresh } from "@material-ui/icons"
+  CardActions,
+} from "@mui/material"
+import { Delete, Refresh, Search } from "@mui/icons-material"
+import SortIcon from "@mui/icons-material/Sort"
+import HomeIcon from "@mui/icons-material/Home"
+import AddBoxIcon from "@mui/icons-material/AddBox"
+import CompareIcon from "@mui/icons-material/Compare"
 
 import { actionCreator } from "../action"
-import { DataGrid, DataGridColumn } from "./DataGrid"
+import { DebouncedInputTextField } from "./Debounce"
 import { studySummariesState } from "../state"
-import { formatDate } from "../dateUtil"
+import { styled } from "@mui/system"
+import { AppDrawer } from "./AppDrawer"
+import { useCreateStudyDialog } from "./CreateStudyDialog"
+import { useDeleteStudyDialog } from "./DeleteStudyDialog"
 
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    card: {
-      margin: theme.spacing(2),
-    },
-    grow: {
-      flexGrow: 1,
-    },
-  })
-)
-
-export const StudyList: FC<{}> = () => {
-  const classes = useStyles()
-  const [openNewStudyDialog, setOpenNewStudyDialog] = React.useState(false)
-  const [openDeleteStudyDialog, setOpenDeleteStudyDialog] = React.useState(
-    false
-  )
-  const [deleteStudyID, setDeleteStudyID] = React.useState(-1)
-  const [newStudyName, setNewStudyName] = React.useState("")
-  const [maximize, setMaximize] = React.useState<boolean>(false)
-
+export const StudyList: FC<{
+  toggleColorMode: () => void
+}> = ({ toggleColorMode }) => {
+  const theme = useTheme()
   const action = actionCreator()
-  const studies = useRecoilValue<StudySummary[]>(studySummariesState)
 
-  const newStudyNameAlreadyUsed = studies.some(
-    (v) => v.study_name === newStudyName
-  )
+  const [studyFilterText, setStudyFilterText] = React.useState<string>("")
+  const studyFilter = (row: StudySummary) => {
+    const keywords = studyFilterText.split(" ")
+    return !keywords.every((k) => {
+      if (k === "") {
+        return true
+      }
+      return row.study_name.indexOf(k) >= 0
+    })
+  }
+  const studies = useRecoilValue<StudySummary[]>(studySummariesState)
+  const [openCreateStudyDialog, renderCreateStudyDialog] =
+    useCreateStudyDialog()
+  const [openDeleteStudyDialog, renderDeleteStudyDialog] =
+    useDeleteStudyDialog()
+
+  const navigate = useNavigate()
+  const useQuery = (): URLSearchParams => {
+    const { search } = useLocation()
+    return useMemo(() => new URLSearchParams(search), [search])
+  }
+  const query = useQuery()
+  const initialSortBy =
+    query.get("studies_order_by") === "desc" ? "desc" : "asc"
+  const [sortBy, setSortBy] = useState<"asc" | "desc">(initialSortBy)
+
+  let filteredStudies = studies.filter((s) => !studyFilter(s))
+
+  if (sortBy === "desc") {
+    filteredStudies = filteredStudies.reverse()
+  }
 
   useEffect(() => {
     action.updateStudySummaries()
   }, [])
 
-  const columns: DataGridColumn<StudySummary>[] = [
-    {
-      field: "study_id",
-      label: "Study ID",
-      sortable: true,
-    },
-    {
-      field: "study_name",
-      label: "Name",
-      sortable: true,
-      toCellValue: (i) => (
-        <Link to={`${URL_PREFIX}/studies/${studies[i].study_id}`}>
-          {studies[i].study_name}
-        </Link>
-      ),
-    },
-    {
-      field: "direction",
-      label: "Direction",
-      sortable: false,
-      toCellValue: (i) => studies[i].direction.toString(),
-    },
-    {
-      field: "best_trial",
-      label: "Best value",
-      sortable: false,
-      toCellValue: (i) => studies[i].best_trial?.value || null,
-    },
-    {
-      field: "datetime_start",
-      label: "Datetime start",
-      sortable: false,
-      toCellValue: (i) => {
-        const datetime_start = studies[i].datetime_start
-        if (datetime_start) {
-          return formatDate(datetime_start)
-        }
-        return null
-      },
-    },
-    {
-      field: "study_name",
-      label: "",
-      sortable: false,
-      padding: "none",
-      toCellValue: (i) => (
-        <IconButton
-          aria-label="delete study"
-          size="small"
-          color="inherit"
-          onClick={(e) => {
-            setDeleteStudyID(studies[i].study_id)
-            setOpenDeleteStudyDialog(true)
-          }}
-        >
-          <Delete />
-        </IconButton>
-      ),
-    },
-  ]
+  useEffect(() => {
+    query.set("studies_order_by", sortBy)
+    navigate(`${location.pathname}?${query.toString()}`, {
+      replace: true,
+    })
+  }, [sortBy])
 
-  const collapseAttrColumns: DataGridColumn<Attribute>[] = [
-    { field: "key", label: "Key", sortable: true },
-    { field: "value", label: "Value", sortable: true },
-  ]
+  const Select = styled(TextField)(({ theme }) => ({
+    "& .MuiInputBase-input": {
+      // vertical padding + font size from searchIcon
+      paddingLeft: `calc(1em + ${theme.spacing(4)})`,
+    },
+  }))
+  const sortBySelect = (
+    <Box
+      sx={{
+        position: "relative",
+        borderRadius: theme.shape.borderRadius,
+        margin: theme.spacing(0, 2),
+      }}
+    >
+      <Box
+        sx={{
+          padding: theme.spacing(0, 2),
+          height: "100%",
+          position: "absolute",
+          pointerEvents: "none",
+          display: "flex",
+          alignItems: "center",
+          justifyContent: "center",
+        }}
+      >
+        <SortIcon />
+      </Box>
+      <Select
+        select
+        value={sortBy}
+        onChange={(e) => {
+          setSortBy(e.target.value as "asc" | "desc")
+        }}
+      >
+        <MenuItem value={"asc"}>Sort ascending</MenuItem>
+        <MenuItem value={"desc"}>Sort descending</MenuItem>
+      </Select>
+    </Box>
+  )
 
-  const handleCloseNewStudyDialog = () => {
-    setNewStudyName("")
-    setOpenNewStudyDialog(false)
-  }
-
-  const handleCreateNewStudy = () => {
-    const direction = maximize ? "maximize" : "minimize"
-    action.createNewStudy(newStudyName, direction)
-    setOpenNewStudyDialog(false)
-    setNewStudyName("")
-  }
-
-  const handleCloseDeleteStudyDialog = () => {
-    setOpenDeleteStudyDialog(false)
-    setDeleteStudyID(-1)
-  }
-
-  const handleDeleteStudy = () => {
-    action.deleteStudy(deleteStudyID)
-    setOpenDeleteStudyDialog(false)
-    setDeleteStudyID(-1)
-  }
-
-  const collapseBody = (index: number) => {
-    return (
-      <Grid container direction="row">
-        <Grid item xs={6}>
-          <Box margin={1}>
-            <Typography variant="h6" gutterBottom component="div">
-              Study user attributes
-            </Typography>
-            <DataGrid<Attribute>
-              columns={collapseAttrColumns}
-              rows={studies[index].user_attrs}
-              keyField={"key"}
-              dense={true}
-              initialRowsPerPage={5}
-              rowsPerPageOption={[5, 10, { label: "All", value: -1 }]}
-            />
-          </Box>
-        </Grid>
-        <Grid item xs={6}>
-          <Box margin={1}>
-            <Typography variant="h6" gutterBottom component="div">
-              Study system attributes
-            </Typography>
-            <DataGrid<Attribute>
-              columns={collapseAttrColumns}
-              rows={studies[index].system_attrs}
-              keyField={"key"}
-              dense={true}
-              initialRowsPerPage={5}
-              rowsPerPageOption={[5, 10, { label: "All", value: -1 }]}
-            />
-          </Box>
-        </Grid>
-      </Grid>
-    )
-  }
+  const toolbar = <HomeIcon sx={{ margin: theme.spacing(0, 1) }} />
 
   return (
-    <div>
-      <AppBar position="static">
-        <Container>
-          <Toolbar>
-            <Typography variant="h6">{APP_BAR_TITLE}</Typography>
-            <div className={classes.grow} />
-            <IconButton
-              aria-controls="menu-appbar"
-              aria-haspopup="true"
-              onClick={(e) => {
-                action.updateStudySummaries("Success to reload")
-              }}
-              color="inherit"
-            >
-              <Refresh />
-            </IconButton>
-            <IconButton
-              aria-controls="menu-appbar"
-              aria-haspopup="true"
-              onClick={(e) => {
-                setOpenNewStudyDialog(true)
-              }}
-              color="inherit"
-            >
-              <AddBox />
-            </IconButton>
-          </Toolbar>
+    <Box sx={{ display: "flex" }}>
+      <AppDrawer toggleColorMode={toggleColorMode} toolbar={toolbar}>
+        <Container
+          sx={{
+            ["@media (min-width: 1280px)"]: {
+              maxWidth: "100%",
+            },
+          }}
+        >
+          <Card sx={{ margin: theme.spacing(2) }}>
+            <CardContent>
+              <Box sx={{ display: "flex" }}>
+                <DebouncedInputTextField
+                  onChange={(s) => {
+                    setStudyFilterText(s)
+                  }}
+                  delay={500}
+                  textFieldProps={{
+                    fullWidth: true,
+                    id: "search-study",
+                    variant: "outlined",
+                    placeholder: "Search study",
+                    sx: { maxWidth: 500 },
+                    InputProps: {
+                      startAdornment: (
+                        <InputAdornment position="start">
+                          <SvgIcon fontSize="small" color="action">
+                            <Search />
+                          </SvgIcon>
+                        </InputAdornment>
+                      ),
+                    },
+                  }}
+                />
+                {sortBySelect}
+                <Box sx={{ flexGrow: 1 }} />
+                <Button
+                  variant="outlined"
+                  startIcon={<Refresh />}
+                  onClick={() => {
+                    action.updateStudySummaries("Success to reload")
+                  }}
+                  sx={{ marginRight: theme.spacing(2), minWidth: "120px" }}
+                >
+                  Reload
+                </Button>
+                <Button
+                  variant="outlined"
+                  startIcon={<AddBoxIcon />}
+                  onClick={() => {
+                    openCreateStudyDialog()
+                  }}
+                  sx={{ marginRight: theme.spacing(2), minWidth: "120px" }}
+                >
+                  Create
+                </Button>
+                <Button
+                  variant="outlined"
+                  startIcon={<CompareIcon />}
+                  component={Link}
+                  to={`${URL_PREFIX}/compare-studies`}
+                  sx={{ marginRight: theme.spacing(2), minWidth: "120px" }}
+                >
+                  Compare
+                </Button>
+              </Box>
+            </CardContent>
+          </Card>
+          <Box sx={{ display: "flex", flexWrap: "wrap" }}>
+            {filteredStudies.map((study) => (
+              <Card
+                key={study.study_id}
+                sx={{ margin: theme.spacing(2), width: "500px" }}
+              >
+                <CardActionArea
+                  component={Link}
+                  to={`${URL_PREFIX}/studies/${study.study_id}`}
+                >
+                  <CardContent>
+                    <Typography variant="h5">
+                      {study.study_id}. {study.study_name}
+                    </Typography>
+                    <Typography
+                      variant="subtitle1"
+                      color="text.secondary"
+                      component="div"
+                    >
+                      {`Direction: ${study.direction.toString()}`}
+                    </Typography>
+                  </CardContent>
+                </CardActionArea>
+                <CardActions disableSpacing sx={{ paddingTop: 0 }}>
+                  <Box sx={{ flexGrow: 1 }} />
+                  <IconButton
+                    aria-label="delete study"
+                    size="small"
+                    color="inherit"
+                    onClick={() => {
+                      openDeleteStudyDialog(study.study_id)
+                    }}
+                  >
+                    <Delete />
+                  </IconButton>
+                </CardActions>
+              </Card>
+            ))}
+          </Box>
         </Container>
-      </AppBar>
-      <Container>
-        <Card className={classes.card}>
-          <DataGrid<StudySummary>
-            columns={columns}
-            rows={studies}
-            keyField={"study_id"}
-            collapseBody={collapseBody}
-            initialRowsPerPage={-1}
-            rowsPerPageOption={[5, 10, { label: "All", value: -1 }]}
-          />
-        </Card>
-      </Container>
-      <Dialog
-        open={openNewStudyDialog}
-        onClose={(e) => {
-          handleCloseNewStudyDialog()
-        }}
-        aria-labelledby="create-study-form-dialog-title"
-      >
-        <DialogTitle id="create-study-form-dialog-title">New study</DialogTitle>
-        <DialogContent>
-          <DialogContentText>
-            To create a new study, please enter the study name here.
-          </DialogContentText>
-          <TextField
-            autoFocus
-            error={newStudyNameAlreadyUsed}
-            helperText={
-              newStudyNameAlreadyUsed ? `"${newStudyName}" is already used` : ""
-            }
-            label="Study name"
-            type="text"
-            onChange={(e) => {
-              setNewStudyName(e.target.value)
-            }}
-            fullWidth
-          />
-          <FormControlLabel
-            control={
-              <Checkbox
-                checked={maximize}
-                onChange={(e) => {
-                  setMaximize(!maximize)
-                }}
-                color="primary"
-              />
-            }
-            label="Set maximize direction"
-          />
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleCloseNewStudyDialog} color="primary">
-            Cancel
-          </Button>
-          <Button
-            onClick={handleCreateNewStudy}
-            color="primary"
-            disabled={newStudyName === "" || newStudyNameAlreadyUsed}
-          >
-            Create
-          </Button>
-        </DialogActions>
-      </Dialog>
-      <Dialog
-        open={openDeleteStudyDialog}
-        onClose={(e) => {
-          handleCloseDeleteStudyDialog()
-        }}
-        aria-labelledby="delete-study-dialog-title"
-      >
-        <DialogTitle id="delete-study-dialog-title">Delete study</DialogTitle>
-        <DialogContent>
-          <DialogContentText>
-            Are you sure you want to delete a study (id={deleteStudyID})?
-          </DialogContentText>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleCloseDeleteStudyDialog} color="primary">
-            No
-          </Button>
-          <Button onClick={handleDeleteStudy} color="primary">
-            Yes
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </div>
+      </AppDrawer>
+      {renderCreateStudyDialog()}
+      {renderDeleteStudyDialog()}
+    </Box>
   )
 }
