@@ -3,6 +3,25 @@ const webpack = require('webpack');
 const mode = process.env.NODE_ENV === 'production' ? 'production' : 'development';
 const isDev = mode === 'development';
 
+const typeScriptLoader = process.env.TYPESCRIPT_LOADER === "esbuild-loader" ? {
+    test: /\.tsx?$/,
+    exclude: [/node_modules/],
+    loader: 'esbuild-loader',
+    options: {
+        loader: 'tsx',
+        tsconfigRaw: require('./tsconfig.json')
+    }
+} : {
+    test: /\.tsx?$/,
+    exclude: [/node_modules/],
+    loader: 'ts-loader',
+    options: {
+        configFile: __dirname + '/tsconfig.json',
+        transpileOnly: isDev,
+        happyPackMode: true
+    }
+}
+
 var config = {
     mode,
     entry: [__dirname + '/src/index.tsx'],
@@ -13,23 +32,14 @@ var config = {
     },
     module: {
         rules: [{
-            oneOf: [
-                {
-                    test: /\.tsx?$/,
-                    exclude: [/node_modules/],
-                    loader: 'ts-loader',
-                    options: {
-                        configFile: __dirname + '/tsconfig.json'
-                    }
-                }
-            ]}]
+            oneOf: [typeScriptLoader]
+        }]
     },
     resolve: {
         extensions: ['.ts', '.tsx', '.js']
     },
     plugins: [
         new webpack.DefinePlugin({
-            'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV),
             'APP_BAR_TITLE': JSON.stringify(process.env.APP_BAR_TITLE || "Goptuna Dashboard"),
             'API_ENDPOINT': JSON.stringify(process.env.API_ENDPOINT),
             'URL_PREFIX': JSON.stringify(process.env.URL_PREFIX || "/dashboard")
@@ -39,9 +49,18 @@ var config = {
 
 if (isDev) {
     config.devtool = 'source-map';
+    config.cache = {
+        type: 'filesystem',
+        buildDependencies: {
+            config: [__filename],
+        }
+    }
     console.log('= = = = = = = = = = = = = = = = = = =');
     console.log('DEVELOPMENT BUILD');
     console.log('= = = = = = = = = = = = = = = = = = =');
+} else {
+    const CompressionPlugin = require("compression-webpack-plugin");
+    config.plugins.push(new CompressionPlugin())
 }
 
 module.exports = config;
